+++ kotlin-inject-compiler/core/src/main/kotlin/me/tatarka/inject/compiler/TypeResultResolver.kt
@@ -25,10 +25,10 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
      * Resolves all [TypeResult] for provider methods in the given class.
      */
     fun resolveAll(context: Context, astClass: AstClass): List<TypeResult.Provider> {
-        return context.types.providerMethods.map { method ->
+        return context.types.providerMembers.map { member ->
             // reset the name allocator between methods so that arg names are not unique across
             // all functions of a component
-            Provider(context.copyNameAllocator(), astClass, method)
+            Provider(context.copyNameAllocator(), astClass, member.member, member.qualifier)
         }
     }
 
@@ -91,7 +91,7 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
         var assistedFailed = false
         val args = context.args.toMutableList()
         for (param in params) {
-            val key = TypeKey(param.type, param.qualifier(options))
+            val key = TypeKey(param.type, param.qualifier(provider, options))
             if (param.isAssisted()) {
                 val arg = args.removeFirstOrNull()
                 if (arg != null) {
@@ -148,7 +148,7 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
         val resolvedImplicitly = mutableListOf<AstParam>()
         for ((i, param) in params.withIndex()) {
             val indexFromEnd = size - i - 1
-            val key = TypeKey(param.type, param.qualifier(options))
+            val key = TypeKey(param.type, param.qualifier(provider, options))
             val arg = args.getOrNull(indexFromEnd)
             if (arg != null) {
                 val (type, name) = arg
@@ -196,11 +196,11 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
 
         val providerResult = types.providerType(key)
         if (providerResult != null) {
-            val (method, types) = providerResult
+            val (member, types) = providerResult
             return Provides(
                 context = withTypes(types),
-                accessor = method.accessor,
-                method = method.method,
+                accessor = member.accessor,
+                method = member.method,
                 scope = null,
                 key = key,
             )
@@ -244,7 +244,7 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
         return null
     }
 
-    private fun Context.method(key: TypeKey, creator: Method): TypeResult {
+    private fun Context.method(key: TypeKey, creator: Member): TypeResult {
         return if (creator.scopedComponent != null && skipScoped != creator.method.returnType) {
             Scoped(
                 context = this,
@@ -392,9 +392,10 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
         context: Context,
         astClass: AstClass,
         method: AstMember,
+        qualifier: AstAnnotation?,
     ): TypeResult.Provider {
         val returnType = method.returnTypeFor(astClass)
-        val key = TypeKey(returnType, method.qualifier(options))
+        val key = TypeKey(returnType, qualifier)
         val result = withCycleDetection(key, method) {
             resolve(context.withoutProvider(returnType), method, key).result
         }
@@ -427,7 +428,7 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
                 methodName = method.name,
                 accessor = accessor,
                 receiver = method.receiverParameterType?.let {
-                    val key = TypeKey(it, method.qualifier(options))
+                    val key = TypeKey(it, key.qualifier)
                     resolve(context, method, key)
                 },
                 isProperty = method is AstProperty,
@@ -468,13 +469,13 @@ class TypeResultResolver(private val provider: AstProvider, private val options:
 
     private fun Container(
         creator: String,
-        args: List<Pair<Method, TypeCollector.Result>>,
-        mapArg: (TypeKey, Method, TypeCollector.Result) -> TypeResult,
+        args: List<Pair<Member, TypeCollector.Result>>,
+        mapArg: (TypeKey, Member, TypeCollector.Result) -> TypeResult,
     ): TypeResult {
         return TypeResult.Container(
             creator = creator,
             args = args.map { (arg, types) ->
-                val key = TypeKey(arg.method.returnType, arg.method.qualifier(options))
+                val key = TypeKey(arg.method.returnType, arg.method.qualifier(provider, options))
                 TypeResultRef(key, mapArg(key, arg, types))
             }
         )